# Fix override existing join types in the query in the `where.associated` method