@p8 2024-02-15
Thanks for the PR @saleh-alhaddad.
Could you explain the use-case this is trying to solve?
Don't `#associated` and `#missing` already cover all cases?
```ruby
class Post < ActiveRecord::Base
  belongs_to :author
end

class Author < ActiveRecord::Base
  has_many :posts
end

author1 = Author.create! name: 'Nora'
post1 = Post.create! title: "First post", author: author1
post2 = Post.create! title: "Second post"
assert_equal Post.where.associated(:author).to_a, [post1]
assert_equal Post.where.missing(:author).to_a, [post2]
```