@saleh-alhaddad 2024-02-16
welcome @p8, my PR is to allow change join type, the `#associated` method uses `joins`, but If I need to change the type to `left_joins` or left_outer_joins, this PR will allow changing the join type from `joins` to any one of allowed `left_joins and left_outer_joins`.