@saleh-alhaddad 2024-02-17
@p8 Thank you for your comments. Let me illustrate the issue with an example:

I need to retrieve all notifications that are associated with a user, including those that are not associated. To accomplish this, I utilize `left_joins`, when I apply the `where.associated` method, the resulting query employs an `inner join`, which excludes cases where there's no association. This behavior conflicts with my requirements.

Initially, I used the following approach:

```ruby
User.left_outer_joins(:comments, :notifications).where.associated(:notifications)
```

which generates the SQL query:


```sql
SELECT * FROM "users" INNER JOIN "notifications" ON "notifications"."user_id" = "users"."id" LEFT OUTER JOIN "comments" ON "comments"."user_id" = "users"."id" WHERE "notifications"."id" IS NOT NULL
```

However, this results in the `where.associated` overriding the `left_joins`, as it employs an inner join, disregarding my intention.

To address this, I need to revert to using `where.associated` and use this way:

```ruby
User.left_joins(:comments, :notifications).where(notifications: { id: nil })
```

This produces the desired SQL query:

```sql

SELECT * FROM "users" LEFT OUTER JOIN "comments" ON "comments"."user_id" = "users"."id" LEFT OUTER JOIN "notifications" ON "notifications"."user_id" = "users"."id" WHERE "notifications"."id" IS NULL
```

This query ensures that both associated and non-associated notifications are included.

Now, to ensure consistency and maintain the appropriate join type, we need to determine a solution. One option is to pass the join type explicitly. Alternatively, we could retain the previously used join type associated with the model, if available. What approach do you suggest?

best regards